+++ src/Site.hs
@@ -1,24 +1,18 @@
-{-# LANGUAGE OverloadedStrings #-}
 module Site where
 
 import Snap
 import Snap.Snaplet.Heist
 import Snap.Snaplet.PostgresqlSimple
-import Data.ByteString
-
+import Database.PostgreSQL.Simple
+import Control.Monad.IO.Class
 import Application
-import User.Handler
+import Routes
+import DataBaseInfo
 
 appInit :: SnapletInit App App
 appInit = makeSnaplet "snap" "SNAP Framework example" Nothing $ do
   hs <- nestSnaplet "heist" heist $ heistInit "templates"
-  p <- nestSnaplet "db" db pgsInit
+  c <- liftIO connInfo
+  p <- nestSnaplet "db" db $ pgsInit' $ (pgsDefaultConfig . postgreSQLConnectionString) c
   addRoutes routes
   return $ App hs p
-
-routes :: [(ByteString,AppHandler)]
-routes = [ ("/", welcomePage)
-         , ("/users", handleUsers)]
-
-welcomePage :: AppHandler
-welcomePage = render "welcome"