+++ src/User/Persistence.hs
@@ -1,7 +1,7 @@
 module User.Persistence where
 
 import Snap.Snaplet.PostgresqlSimple
-import User.Query
+import User.Queries
 import User.Types
 
 selectAllUsers :: (HasPostgres m) => m [User]