+++ src/User/Query.hs
@@ -1,9 +1,0 @@
-{-# LANGUAGE QuasiQuotes #-}
-module User.Query where
-
-import Database.PostgreSQL.Simple.SqlQQ
-import Snap.Snaplet.PostgresqlSimple
-
-allUsersQuery :: Query
-allUsersQuery = [sql| SELECT *
-                       FROM snap_user |]