+++ src/User/Types.hs
@@ -1,12 +1,34 @@
 {-# OPTIONS_GHC -fno-warn-orphans #-}
-{-# LANGUAGE RecordWildCards, TypeSynonymInstances, FlexibleInstances, OverloadedStrings #-}
+{-# LANGUAGE TypeSynonymInstances #-}
 module User.Types where
 
 import Data.Time.Calendar
 import Snap.Snaplet.PostgresqlSimple
 import Database.PostgreSQL.Simple.ToField
 import Database.PostgreSQL.Simple.FromField
-import Data.Scientific
+import qualified Data.ByteString.Char8 as B
+import Web.PathPieces
+import qualified Data.Text as T
+import Text.Read
+import Text.Printf
+
+newtype Currency 
+  = Currency 
+  { unCurrency :: Double } 
+  deriving (Show,Eq)
+
+instance ToField Currency where
+  toField = toField . unCurrency
+
+instance FromField Currency where
+  fromField _ md = return $ Currency $
+    case md of
+      Just d -> read $ B.unpack d :: Double
+      _      -> 0
+
+instance PathPiece Currency where
+  fromPathPiece = fmap Currency . readMaybe @Double . T.unpack
+  toPathPiece = T.pack . printf "%.2f" . unCurrency
 
 data User = User
   { userKey       :: Int
@@ -15,7 +37,7 @@ data User = User
   , userBirthDate :: Day
   , userPassword  :: String
   , userGender    :: Gender
-  , userIncome    :: Scientific}
+  , userIncome    :: Currency}
 
 data Gender 
   = Male
@@ -52,3 +74,11 @@ instance ToRow User where
     , toField userPassword
     , toField userGender
     , toField userIncome]
+
+data UserPresenter
+  = UserPresenter
+  { upName     :: String
+  , upMail     :: String
+  , upBirthday :: String
+  , upIncome   :: String}
+  deriving (Show,Eq)