+++ ChangeLog.md
@@ -4,6 +4,7 @@
     - Added a base template.
     - Added the User query, persistence and handler modules.
     - Added the user list template.
+    - Changed init of db snaplet with the loading of .env values.
 
 ## 1.0.3
     - Changed snaplet-persistent for snaplet-postgresql-simple.